req validation

+++ controllers/cards.js
@@ -1,57 +1,55 @@
 import Card from '../models/card';
+import NotFoundError from '../errors/NotFoundError';
+import BadRequestError from '../errors/BadRequestError';
+import ForbiddenError from '../errors/ForbiddenError';
 
-const ERROR_CODE = 400;
-const NOTFOUND_CODE = 404;
-const DEFAULTERROR_CODE = 500;
-
-const errorHandling = (err, res) => {
-  switch (err.name) {
-    case 'CastError':
-    case 'ValidationError':
-      res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные.', err });
-      break;
-    case 'ReferenceError':
-      res.status(NOTFOUND_CODE).send({ message: 'Карточка с указанным _id не найдена.' });
-      break;
-    default:
-      res.status(DEFAULTERROR_CODE).send({ message: 'Произошла ошибка' });
-      break;
-  }
-};
-
-export const getCards = (req, res) => {
+export const getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
-    .then((cards) => res.send(cards))
-    .catch((err) => errorHandling(err, res));
+    .then((cards) => {
+      if (!cards) {
+        throw new NotFoundError('Карточки не найдены');
+      }
+      res.send(cards);
+    })
+    .catch(next);
 };
 
-export const createCards = (req, res) => {
-  // const { _id } = req.user;
+export const createCards = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user })
-    .then((card) => res.send(card))
-    .catch((err) => errorHandling(err, res));
+    .then((card) => {
+      if (!card) {
+        throw new BadRequestError('Переданы некорректные данные');
+      }
+      res.send(card);
+    })
+    .catch(next);
 };
 
-export const deleteCards = (req, res) => {
+export const deleteCards = (req, res, next) => {
   Card.findById({ _id: req.params.cardId })
     .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Карточка не найдена');
+      }
       if (req.user._id === card.owner._id.toString()) {
         Card.findByIdAndRemove({ _id: req.params.cardId })
           .then((data) => {
-            if (data) {
-              res.send({ message: 'Карточка удалена' });
-            } else throw new ReferenceError();
+            if (!data) {
+              throw new ReferenceError();
+            }
+            res.send({ message: 'Карточка удалена' });
           })
-          .catch((err) => errorHandling(err, res));
+          .catch(next);
       } else {
-        res.status(403).send({ message: 'Вы не владелец карточки' });
+        throw new ForbiddenError('Вы не владелец карточки');
       }
-    });
+    })
+    .catch(next);
 };
 
-export const likeCard = (req, res) => {
+export const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -59,14 +57,15 @@ export const likeCard = (req, res) => {
   )
     .populate(['owner', 'likes'])
     .then((card) => {
-      if (card) {
-        res.send(card);
-      } else throw new ReferenceError();
+      if (!card) {
+        throw new NotFoundError('Карточка не найдена');
+      }
+      res.send(card);
     })
-    .catch((err) => errorHandling(err, res));
+    .catch(next);
 };
 
-export const dislikeCard = (req, res) => {
+export const dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -78,5 +77,5 @@ export const dislikeCard = (req, res) => {
         res.send(card);
       } else throw new ReferenceError();
     })
-    .catch((err) => errorHandling(err, res));
+    .catch(next);
 };
